rate limit the requests to github
I would appear that github has implemented a new ddos protection
method that is causing consistent failures and timeouts when making
more than 50 requests in short succession.

This change introduces async and uses its queue functionality to force
a rate limit on the network requests, the result seems to potentially
be an overall faster experience for lower numbers of changes being compared.

Currently async is queuing 15 items. This number could be tuned or
potentially exposed as variables.