fix: make test pass
Ref: https://github.com/nodejs/commit-stream/pull/10

The above change to commit-stream makes it strip out `(#xx)` from the summary
and use it as one of the inputs to find the PR url. The change here just
updates the test assertion to remove that from the commit—we don't need it and
this is an improvement in functionality.